refactor(image-mode): recipe pass + extract composable
